Relocate tunnel services object
Moves the `tunnelServices` object into the `ssh-tunneling` module and deletes the `tunnel-services` module altogether.
