+++ src/commands/borealis-pg/tunnel.test.ts
@@ -12,8 +12,8 @@ import {
   verify,
   when,
 } from 'ts-mockito'
+import {tunnelServices} from '../../ssh-tunneling'
 import {borealisPgApiBaseUrl, expect, herokuApiBaseUrl, test} from '../../test-utils'
-import tunnelServices from '../../tunnel-services'
 
 const localPgHostname = 'pg-tunnel.borealis-data.com'
 const defaultSshPort = 22
